started merging backend/frontend of waterfall
